#56 moved cmd pkg to server
