+++ src/server/cmd/artwork/artwork_extract_handler.go
