+++ src/server/cmd/audioscan/artwork_update_handler.go
