+++ src/server/cmd/audioscan/directory_scan.go
