+++ src/server/cmd/mediainfo/main.go
