+++ src/server/cmd/player/mpv.md
