+++ src/server/cmd/player/socket.sh
