+++ src/server/cmd/restapi/main.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	_ "github.com/lib/pq"
-	"github.com/mauleyzaola/maupod/src/server/maupod/cmd/restapi/pkg/api"
+	"github.com/mauleyzaola/maupod/src/server/cmd/restapi/pkg/api"
 	"github.com/mauleyzaola/maupod/src/server/pkg/dbdata"
 	"github.com/mauleyzaola/maupod/src/server/pkg/helpers"
 	"github.com/mauleyzaola/maupod/src/server/pkg/rules"
