+++ src/server/dockerfiles/artwork.Dockerfile
@@ -12,4 +12,4 @@ RUN ["go", "get", "-v", "github.com/githubnemo/CompileDaemon"]
 
 FROM golang-dev
 WORKDIR /go/src/github.com/mauleyzaola/maupod/src/server
-ENTRYPOINT CompileDaemon -directory=. -log-prefix=false -build="go build -o maupod-artwork ./maupod/cmd/artwork" -command="./maupod-artwork"
+ENTRYPOINT CompileDaemon -directory=. -log-prefix=false -build="go build -o maupod-artwork ./cmd/artwork" -command="./maupod-artwork"
