+++ src/server/dockerfiles/mediainfo.Dockerfile
@@ -10,4 +10,4 @@ RUN ["go", "get", "-v", "github.com/githubnemo/CompileDaemon"]
 
 FROM golang-dev
 WORKDIR /go/src/github.com/mauleyzaola/maupod/src/server
-ENTRYPOINT CompileDaemon -directory=. -log-prefix=false -build="go build -o maupod-mediainfo ./maupod/cmd/mediainfo" -command="./maupod-mediainfo"
+ENTRYPOINT CompileDaemon -directory=. -log-prefix=false -build="go build -o maupod-mediainfo ./cmd/mediainfo" -command="./maupod-mediainfo"
